+++ packages/compiler/src/injectable_compiler_2.ts
@@ -30,6 +30,7 @@ export interface IvyInjectableDep {
   optional: boolean;
   self: boolean;
   skipSelf: boolean;
+  attribute: boolean;
 }
 
 export interface IvyInjectableMetadata {
