+++ packages/compiler/test/render3/r3_view_compiler_input_outputs_spec.ts
@@ -62,6 +62,7 @@ describe('compiler compliance: listen()', () => {
             componentOutput: 'componentOutput',
             originalComponentOutput: 'renamedComponentOutput'
           }
+          …
         });`;
 
     const directiveDef = `
@@ -75,6 +76,7 @@ describe('compiler compliance: listen()', () => {
           directiveOutput: 'directiveOutput',
           originalDirectiveOutput: 'renamedDirectiveOutput'
         }
+        …
       });`;
